requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-core==1.32.0
azure-cosmos==4.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
joblib==1.4.2
numpy==2.1.3
openai==1.55.0
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
